+++ test/suite/NpmBridgeFactoryTest.php
@@ -1,20 +1,11 @@
 <?php
 
-/*
- * This file is part of the Composer NPM bridge package.
- *
- * Copyright © 2016 Erin Millard
- *
- * For the full copyright and license information, please view the LICENSE
- * file that was distributed with this source code.
- */
-
 namespace Eloquent\Composer\NpmBridge;
 
 use Composer\IO\NullIO;
-use PHPUnit_Framework_TestCase;
+use PHPUnit\Framework\TestCase;
 
-class NpmBridgeFactoryTest extends PHPUnit_Framework_TestCase
+class NpmBridgeFactoryTest extends TestCase
 {
     protected function setUp()
     {