+++ test/suite/NpmBridgePluginTest.php
@@ -1,14 +1,5 @@
 <?php
 
-/*
- * This file is part of the Composer NPM bridge package.
- *
- * Copyright © 2016 Erin Millard
- *
- * For the full copyright and license information, please view the LICENSE
- * file that was distributed with this source code.
- */
-
 namespace Eloquent\Composer\NpmBridge;
 
 use Composer\Composer;
@@ -16,20 +7,20 @@ use Composer\IO\NullIO;
 use Composer\Script\Event;
 use Composer\Script\ScriptEvents;
 use Eloquent\Phony\Phpunit\Phony;
-use PHPUnit_Framework_TestCase;
+use PHPUnit\Framework\TestCase;
 
-class NpmBridgePluginTest extends PHPUnit_Framework_TestCase
+class NpmBridgePluginTest extends TestCase
 {
     protected function setUp()
     {
         $this->bridgeFactory = Phony::mock('Eloquent\Composer\NpmBridge\NpmBridgeFactory');
-        $this->plugin = new NpmBridgePlugin($this->bridgeFactory->mock());
+        $this->plugin = new NpmBridgePlugin($this->bridgeFactory->get());
 
         $this->bridge = Phony::mock('Eloquent\Composer\NpmBridge\NpmBridge');
         $this->composer = new Composer();
         $this->io = new NullIO();
 
-        $this->bridgeFactory->createBridge('*')->returns($this->bridge);
+        $this->bridgeFactory->createBridge->returns($this->bridge);
     }
 
     public function testConstructorWithoutArguments()
@@ -45,10 +36,10 @@ class NpmBridgePluginTest extends PHPUnit_Framework_TestCase
     public function testGetSubscribedEvents()
     {
         $this->assertSame(
-            array(
+            [
                 ScriptEvents::POST_INSTALL_CMD => 'onPostInstallCmd',
                 ScriptEvents::POST_UPDATE_CMD => 'onPostUpdateCmd',
-            ),
+            ],
             $this->plugin->getSubscribedEvents()
         );
     }