+++ test/suite/NpmBridgeTest.php
@@ -1,14 +1,5 @@
 <?php
 
-/*
- * This file is part of the Composer NPM bridge package.
- *
- * Copyright © 2016 Erin Millard
- *
- * For the full copyright and license information, please view the LICENSE
- * file that was distributed with this source code.
- */
-
 namespace Eloquent\Composer\NpmBridge;
 
 use Composer\Composer;
@@ -16,16 +7,16 @@ use Composer\Package\Link;
 use Composer\Package\Package;
 use Composer\Package\RootPackage;
 use Eloquent\Phony\Phpunit\Phony;
-use PHPUnit_Framework_TestCase;
+use PHPUnit\Framework\TestCase;
 
-class NpmBridgeTest extends PHPUnit_Framework_TestCase
+class NpmBridgeTest extends TestCase
 {
     protected function setUp()
     {
         $this->io = Phony::mock('Composer\IO\IOInterface');
         $this->vendorFinder = Phony::mock('Eloquent\Composer\NpmBridge\NpmVendorFinder');
         $this->client = Phony::mock('Eloquent\Composer\NpmBridge\NpmClient');
-        $this->bridge = new NpmBridge($this->io->mock(), $this->vendorFinder->mock(), $this->client->mock());
+        $this->bridge = new NpmBridge($this->io->get(), $this->vendorFinder->get(), $this->client->get());
 
         $this->composer = new Composer();
 
@@ -38,17 +29,17 @@ class NpmBridgeTest extends PHPUnit_Framework_TestCase
         $this->linkRoot3 = new Link('vendor/package', 'eloquent/composer-npm-bridge');
 
         $this->installationManager = Phony::mock('Composer\Installer\InstallationManager');
-        $this->installationManager->getInstallPath($this->packageA)->returns('/path/to/install/a');
-        $this->installationManager->getInstallPath($this->packageB)->returns('/path/to/install/b');
+        $this->installationManager->getInstallPath->with($this->packageA)->returns('/path/to/install/a');
+        $this->installationManager->getInstallPath->with($this->packageB)->returns('/path/to/install/b');
 
         $this->composer->setPackage($this->rootPackage);
-        $this->composer->setInstallationManager($this->installationManager->mock());
+        $this->composer->setInstallationManager($this->installationManager->get());
     }
 
     public function testInstall()
     {
-        $this->rootPackage->setRequires(array($this->linkRoot1, $this->linkRoot2, $this->linkRoot3));
-        $this->vendorFinder->find($this->composer, $this->bridge)->returns(array($this->packageA, $this->packageB));
+        $this->rootPackage->setRequires([$this->linkRoot1, $this->linkRoot2, $this->linkRoot3]);
+        $this->vendorFinder->find->with($this->composer, $this->bridge)->returns([$this->packageA, $this->packageB]);
         $this->bridge->install($this->composer);
 
         Phony::inOrder(
@@ -64,8 +55,8 @@ class NpmBridgeTest extends PHPUnit_Framework_TestCase
 
     public function testInstallProductionMode()
     {
-        $this->rootPackage->setRequires(array($this->linkRoot1, $this->linkRoot2, $this->linkRoot3));
-        $this->vendorFinder->find($this->composer, $this->bridge)->returns(array($this->packageA, $this->packageB));
+        $this->rootPackage->setRequires([$this->linkRoot1, $this->linkRoot2, $this->linkRoot3]);
+        $this->vendorFinder->find->with($this->composer, $this->bridge)->returns([$this->packageA, $this->packageB]);
         $this->bridge->install($this->composer, false);
 
         Phony::inOrder(
@@ -81,8 +72,8 @@ class NpmBridgeTest extends PHPUnit_Framework_TestCase
 
     public function testInstallRootDevDependenciesInDevMode()
     {
-        $this->rootPackage->setDevRequires(array($this->linkRoot3));
-        $this->vendorFinder->find($this->composer, $this->bridge)->returns(array());
+        $this->rootPackage->setDevRequires([$this->linkRoot3]);
+        $this->vendorFinder->find->with($this->composer, $this->bridge)->returns([]);
         $this->bridge->install($this->composer, true);
 
         $this->client->install->calledWith(null, true);
@@ -90,8 +81,8 @@ class NpmBridgeTest extends PHPUnit_Framework_TestCase
 
     public function testInstallRootDevDependenciesInProductionMode()
     {
-        $this->rootPackage->setDevRequires(array($this->linkRoot3));
-        $this->vendorFinder->find($this->composer, $this->bridge)->returns(array());
+        $this->rootPackage->setDevRequires([$this->linkRoot3]);
+        $this->vendorFinder->find->with($this->composer, $this->bridge)->returns([]);
         $this->bridge->install($this->composer, false);
 
         $this->client->install->never()->called();
@@ -99,8 +90,8 @@ class NpmBridgeTest extends PHPUnit_Framework_TestCase
 
     public function testInstallNothing()
     {
-        $this->rootPackage->setRequires(array($this->linkRoot1, $this->linkRoot2));
-        $this->vendorFinder->find($this->composer, $this->bridge)->returns(array());
+        $this->rootPackage->setRequires([$this->linkRoot1, $this->linkRoot2]);
+        $this->vendorFinder->find->with($this->composer, $this->bridge)->returns([]);
         $this->bridge->install($this->composer);
 
         Phony::inOrder(
@@ -113,15 +104,14 @@ class NpmBridgeTest extends PHPUnit_Framework_TestCase
 
     public function testUpdate()
     {
-        $this->rootPackage->setRequires(array($this->linkRoot1, $this->linkRoot2, $this->linkRoot3));
-        $this->vendorFinder->find($this->composer, $this->bridge)->returns(array($this->packageA, $this->packageB));
+        $this->rootPackage->setRequires([$this->linkRoot1, $this->linkRoot2, $this->linkRoot3]);
+        $this->vendorFinder->find->with($this->composer, $this->bridge)->returns([$this->packageA, $this->packageB]);
         $this->bridge->update($this->composer);
 
         Phony::inOrder(
             $this->io->write->calledWith('<info>Updating NPM dependencies for root project</info>'),
             $this->client->update->calledWith(),
             $this->client->install->calledWith(null, true),
-            $this->client->shrinkwrap->calledWith(),
             $this->io->write->calledWith('<info>Installing NPM dependencies for Composer dependencies</info>'),
             $this->io->write->calledWith('<info>Installing NPM dependencies for vendorA/packageA</info>'),
             $this->client->install->calledWith('/path/to/install/a', false),
@@ -132,8 +122,8 @@ class NpmBridgeTest extends PHPUnit_Framework_TestCase
 
     public function testUpdateNothing()
     {
-        $this->rootPackage->setRequires(array($this->linkRoot1, $this->linkRoot2));
-        $this->vendorFinder->find($this->composer, $this->bridge)->returns(array());
+        $this->rootPackage->setRequires([$this->linkRoot1, $this->linkRoot2]);
+        $this->vendorFinder->find->with($this->composer, $this->bridge)->returns([]);
         $this->bridge->update($this->composer);
 
         Phony::inOrder(
@@ -146,8 +136,8 @@ class NpmBridgeTest extends PHPUnit_Framework_TestCase
 
     public function testIsDependantPackage()
     {
-        $this->packageA->setRequires(array($this->linkRoot3));
-        $this->packageB->setDevRequires(array($this->linkRoot3));
+        $this->packageA->setRequires([$this->linkRoot3]);
+        $this->packageB->setDevRequires([$this->linkRoot3]);
 
         $this->assertTrue($this->bridge->isDependantPackage($this->packageA));
         $this->assertFalse($this->bridge->isDependantPackage($this->packageB));