+++ test/suite/NpmVendorFinderTest.php
@@ -1,14 +1,5 @@
 <?php
 
-/*
- * This file is part of the Composer NPM bridge package.
- *
- * Copyright © 2016 Erin Millard
- *
- * For the full copyright and license information, please view the LICENSE
- * file that was distributed with this source code.
- */
-
 namespace Eloquent\Composer\NpmBridge;
 
 use Composer\Composer;
@@ -17,9 +8,9 @@ use Composer\Package\Link;
 use Composer\Package\Package;
 use Composer\Repository\ArrayRepository;
 use Eloquent\Phony\Phpunit\Phony;
-use PHPUnit_Framework_TestCase;
+use PHPUnit\Framework\TestCase;
 
-class NpmVendorFinderTest extends PHPUnit_Framework_TestCase
+class NpmVendorFinderTest extends TestCase
 {
     protected function setUp()
     {
@@ -45,7 +36,7 @@ class NpmVendorFinderTest extends PHPUnit_Framework_TestCase
         $this->linkD1 = new Link('vendorD/packageD', 'eloquent/composer-npm-bridge');
         $this->linkD2 = new Link('vendorD/packageD', 'vendorZ/packageZ');
 
-        $this->composer->setRepositoryManager($this->repositoryManager->mock());
+        $this->composer->setRepositoryManager($this->repositoryManager->get());
         $this->repositoryManager->getLocalRepository->returns($this->localRepository);
 
         $this->localRepository->addPackage($this->packageA);
@@ -53,14 +44,14 @@ class NpmVendorFinderTest extends PHPUnit_Framework_TestCase
         $this->localRepository->addPackage($this->packageC);
         $this->localRepository->addPackage($this->packageD);
 
-        $this->packageA->setRequires(array($this->linkA1, $this->linkA2));
-        $this->packageB->setRequires(array($this->linkB1, $this->linkB2));
-        $this->packageC->setDevRequires(array($this->linkC1, $this->linkC2));
-        $this->packageD->setRequires(array($this->linkD1, $this->linkD2));
+        $this->packageA->setRequires([$this->linkA1, $this->linkA2]);
+        $this->packageB->setRequires([$this->linkB1, $this->linkB2]);
+        $this->packageC->setDevRequires([$this->linkC1, $this->linkC2]);
+        $this->packageD->setRequires([$this->linkD1, $this->linkD2]);
     }
 
     public function testFind()
     {
-        $this->assertSame(array($this->packageB, $this->packageD), $this->finder->find($this->composer, $this->bridge));
+        $this->assertSame([$this->packageB, $this->packageD], $this->finder->find($this->composer, $this->bridge));
     }
 }